requirements.txt
alembic==1.8.1
certifi==2022.9.24
charset-normalizer==2.1.1
greenlet==2.0.1
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.44
urllib3==1.26.13
